Enable autoGc of WeakConcurrentBag on JVM only
