+++ core/js/src/main/scala/zio/internal/WeakConcurrentBagGc.scala
@@ -1,0 +1,7 @@
+package zio.internal
+
+import zio.Duration
+
+private object WeakConcurrentBagGc {
+  def start[A <: AnyRef](bag: WeakConcurrentBag[A], every: Duration): Unit = ()
+}
